+++ src/chart_types/xy_chart/rendering/rendering.ts
@@ -1,4 +1,5 @@
 import { area, line } from 'd3-shape';
+import { $Values } from 'utility-types';
 
 import { CanvasTextBBoxCalculator } from '../../../utils/bbox/canvas_text_bbox_calculator';
 import {
@@ -7,8 +8,9 @@ import {
   LineSeriesStyle,
   LineStyle,
   PointStyle,
-  SharedGeometryStyle,
+  SharedGeometryStateStyle,
   BarSeriesStyle,
+  GeometryStateStyle,
 } from '../../../utils/themes/theme';
 import { SpecId } from '../../../utils/ids';
 import { isLogarithmicScale } from '../../../utils/scales/scale_continuous';
@@ -33,7 +35,7 @@ export const AccessorType = Object.freeze({
   Y1: 'y1' as 'y1',
 });
 
-export type AccessorType = typeof AccessorType.Y0 | typeof AccessorType.Y1;
+export type AccessorType = $Values<typeof AccessorType>;
 
 export interface GeometryValue {
   y: any;
@@ -41,17 +43,14 @@ export interface GeometryValue {
   accessor: AccessorType;
 }
 
-/** Shared style properties for varies geometries */
-export interface GeometryStyle {
-  /**
-   * Opacity multiplier
-   *
-   * if set to `0.5` all given opacities will be halfed
-   */
-  opacity: number;
-}
-
 export type IndexedGeometry = PointGeometry | BarGeometry;
+
+/**
+ * Array of **range** clippings [x1, x2] to be excluded during rendering
+ *
+ * Note: Must be scaled **range** values (i.e. pixel coordinates) **NOT** domain values
+ */
+export type ClippedRanges = [number, number][];
 
 export interface PointGeometry {
   x: number;
@@ -94,6 +93,10 @@ export interface LineGeometry {
   geometryId: GeometryId;
   seriesLineStyle: LineStyle;
   seriesPointStyle: PointStyle;
+  /**
+   * Ranges of `[x0, x1]` pairs to clip from series
+   */
+  clippedRanges: ClippedRanges;
 }
 export interface AreaGeometry {
   area: string;
@@ -109,6 +112,10 @@ export interface AreaGeometry {
   seriesAreaLineStyle: LineStyle;
   seriesPointStyle: PointStyle;
   isStacked: boolean;
+  /**
+   * Ranges of `[x0, x1]` pairs to clip from series
+   */
+  clippedRanges: ClippedRanges;
 }
 
 export function isPointGeometry(ig: IndexedGeometry): ig is PointGeometry {
@@ -416,6 +423,7 @@ export function renderLine(
   xScaleOffset: number,
   seriesStyle: LineSeriesStyle,
   pointStyleAccessor?: PointStyleAccessor,
+  hasFit?: boolean,
 ): {
   lineGeometry: LineGeometry;
   indexedGeometries: Map<any, IndexedGeometry[]>;
@@ -424,15 +432,19 @@ export function renderLine(
 
   const pathGenerator = line<DataSeriesDatum>()
     .x(({ x }) => xScale.scale(x) - xScaleOffset)
-    .y(({ y1 }) => {
-      if (y1 !== null) {
-        return yScale.scale(y1);
+    .y((datum) => {
+      const yValue = getYValue(datum);
+
+      if (yValue !== null) {
+        return yScale.scale(yValue);
       }
+
       // this should never happen thanks to the defined function
       return yScale.isInverted ? yScale.range[1] : yScale.range[0];
     })
-    .defined(({ x, y1 }) => {
-      return y1 !== null && !(isLogScale && y1 <= 0) && xScale.isValueInDomain(x);
+    .defined((datum) => {
+      const yValue = getYValue(datum);
+      return yValue !== null && !(isLogScale && yValue <= 0) && xScale.isValueInDomain(datum.x);
     })
     .curve(getCurveFactory(curve));
   const y = 0;
@@ -450,6 +462,7 @@ export function renderLine(
     pointStyleAccessor,
   );
 
+  const clippedRanges = hasFit && !hasY0Accessors ? getClippedRanges(dataset, xScale, xScaleOffset) : [];
   const lineGeometry = {
     line: pathGenerator(dataset) || '',
     points: pointGeometries,
@@ -464,12 +477,28 @@ export function renderLine(
     },
     seriesLineStyle: seriesStyle.line,
     seriesPointStyle: seriesStyle.point,
+    clippedRanges,
   };
   return {
     lineGeometry,
     indexedGeometries,
   };
 }
+
+/**
+ * Returns value of `y1` or `filled.y1` or null
+ */
+export const getYValue = ({ y1, filled }: DataSeriesDatum): number | null => {
+  if (y1 !== null) {
+    return y1;
+  }
+
+  if (filled && filled.y1 !== undefined) {
+    return filled.y1;
+  }
+
+  return null;
+};
 
 export function renderArea(
   shift: number,
@@ -485,17 +514,18 @@ export function renderArea(
   seriesStyle: AreaSeriesStyle,
   isStacked = false,
   pointStyleAccessor?: PointStyleAccessor,
+  hasFit?: boolean,
 ): {
   areaGeometry: AreaGeometry;
   indexedGeometries: Map<any, IndexedGeometry[]>;
 } {
   const isLogScale = isLogarithmicScale(yScale);
-
   const pathGenerator = area<DataSeriesDatum>()
     .x(({ x }) => xScale.scale(x) - xScaleOffset)
-    .y1(({ y1 }) => {
-      if (y1 !== null) {
-        return yScale.scale(y1);
+    .y1((datum) => {
+      const yValue = getYValue(datum);
+      if (yValue !== null) {
+        return yScale.scale(yValue);
       }
       // this should never happen thanks to the defined function
       return yScale.isInverted ? yScale.range[1] : yScale.range[0];
@@ -506,13 +536,14 @@ export function renderArea(
       }
       return yScale.scale(y0);
     })
-    .defined(({ y1, x }) => {
-      return y1 !== null && !(isLogScale && y1 <= 0) && xScale.isValueInDomain(x);
+    .defined((datum) => {
+      const yValue = getYValue(datum);
+      return yValue !== null && !(isLogScale && yValue <= 0) && xScale.isValueInDomain(datum.x);
     })
     .curve(getCurveFactory(curve));
 
+  const clippedRanges = hasFit && !hasY0Accessors && !isStacked ? getClippedRanges(dataset, xScale, xScaleOffset) : [];
   const y1Line = pathGenerator.lineY1()(dataset);
-
   const lines: string[] = [];
   if (y1Line) {
     lines.push(y1Line);
@@ -536,7 +567,7 @@ export function renderArea(
     pointStyleAccessor,
   );
 
-  const areaGeometry = {
+  const areaGeometry: AreaGeometry = {
     area: pathGenerator(dataset) || '',
     lines,
     points: pointGeometries,
@@ -553,6 +584,7 @@ export function renderArea(
     seriesAreaLineStyle: seriesStyle.line,
     seriesPointStyle: seriesStyle.point,
     isStacked,
+    clippedRanges,
   };
   return {
     areaGeometry,
@@ -560,12 +592,47 @@ export function renderArea(
   };
 }
 
-export function getGeometryStyle(
+/**
+ * Gets clipped ranges that have been fitted to values
+ * @param dataset
+ * @param xScale
+ * @param xScaleOffset
+ */
+export function getClippedRanges(dataset: DataSeriesDatum[], xScale: Scale, xScaleOffset: number): ClippedRanges {
+  let firstNonNullX: number | null = null;
+  let hasNull = false;
+
+  return dataset.reduce<ClippedRanges>((acc, { x, y1 }) => {
+    const xValue = xScale.scale(x) - xScaleOffset + xScale.bandwidth / 2;
+
+    if (y1 !== null) {
+      if (hasNull) {
+        if (firstNonNullX !== null) {
+          acc.push([firstNonNullX, xValue]);
+        } else {
+          acc.push([0, xValue]);
+        }
+        hasNull = false;
+      }
+
+      firstNonNullX = xValue;
+    } else {
+      const endXValue = xScale.range[1] - xScale.bandwidth * (2 / 3);
+      if (firstNonNullX !== null && xValue === endXValue) {
+        acc.push([firstNonNullX, xValue]);
+      }
+      hasNull = true;
+    }
+    return acc;
+  }, []);
+}
+
+export function getGeometryStateStyle(
   geometryId: GeometryId,
   highlightedLegendItem: LegendItem | null,
-  sharedGeometryStyle: SharedGeometryStyle,
+  sharedGeometryStyle: SharedGeometryStateStyle,
   individualHighlight?: { [key: string]: boolean },
-): GeometryStyle {
+): GeometryStateStyle {
   const { default: defaultStyles, highlighted, unhighlighted } = sharedGeometryStyle;
 
   if (highlightedLegendItem != null) {